# ActiveRecord initialization optimizations