@maclover7 2017-05-24
If there are other possible performance slowdowns, feel free to open up an issue to identify them, and then we can open them up to community contributions :)